fix(auth): address PR #71 Copilot review (5 findings)
- auth/reserved: return OpError::Forbidden(reason) instead of
  InvalidRequest. Reserved-label rejection is an authorization outcome,
  not a malformed request; Forbidden surfaces as HTTP 403 (+ preserved
  reason) and GQLSTATUS 42501 while InvalidRequest mapped to 400. New
  variant added (unit AuthDenied kept for the reason-less case). Doc
  comments updated to match. Wired into HTTP, QUIC, and MCP error
  mappers.

- ops::principals::encode_scope_root_ids: store NodeIds as Value::UInt
  instead of casting u64 to i64 via `as`. The cast silently wrapped
  for ids above i64::MAX and got filtered out at read time by
  scope_roots (which drops negatives), corrupting scope data for
  principals pointing at high-id nodes. scope_roots reader already
  accepts both Int and UInt so this is backward compatible.

- bootstrap::migrate_main_graph_principals_to_vault: same u64→i64
  wrap bug in the legacy-principal migration path; switched to
  Value::UInt. Without the fix a migrating deployment could silently
  drop scope roots above i64::MAX during 1.3.0 upgrade.

- auth/oauth::validate: scope_generation is now stamped from the
  main graph containment_generation, not the vault generation.
  Previously build_auth_context received vault_gen and stuffed it
  into AuthContext::scope_generation, but ops::refresh_scope_if_stale
  compares that against state.graph.containment_generation() — two
  unrelated counters — so every post-validate op saw "stale" and
  recomputed scope. Performance-only issue (scope was still
  correct), but fixed now. vault_gen stays as the principal_cache
  key where it belongs; doc comment on build_auth_context updated
  to flag the two-generation distinction explicitly.

- ops::refresh_scope_if_stale: the vault-unavailable branch now
  fails closed instead of returning the prior AuthContext unchanged.
  The old behavior preserved whatever scope bitmap the principal
  last carried, which combined with an absent vault could let ops
  succeed after identity was logically detached. Now drops scope to
  empty and stamps the current containment generation, so
  require_in_scope rejects and the next refresh still runs.

Verification
============
- `cargo fmt --all`
- `cargo clippy --workspace --features selene-server/dev-tls,selene-testing/bench --all-targets -- -D warnings`: clean
- `cargo test --workspace --features selene-server/dev-tls,selene-testing/bench`: all green, 349 selene-server lib tests still pass.

Addresses all 4 inline comments + the low-confidence summary item on
PR #71. Part of PR1 (WorkItem 11052).

+++ crates/selene-server/src/bootstrap.rs
@@ -1089,10 +1089,14 @@ fn migrate_main_graph_principals_to_vault(
             props.push((IStr::new("redirect_uri"), Value::str(uri)));
         }
         if !p.scope_roots.is_empty() {
+            // Stored as UInt because NodeId is u64; casting to i64 via `as`
+            // would silently wrap for ids above i64::MAX and the reader
+            // (`projection::scope_roots`) drops negative items, corrupting
+            // migrated scope data.
             let values: Arc<[Value]> = p
                 .scope_roots
                 .iter()
-                .map(|id| Value::Int(*id as i64))
+                .map(|id| Value::UInt(*id))
                 .collect::<Vec<_>>()
                 .into();
             props.push((IStr::new("scope_root_ids"), Value::List(values)));
